requirements.txt
python-dotenv
setuptools
yt_dlp
moviepy
inquirer